requirements.txt
discord.py>=2.3.0
gspread>=5.12.0
google-auth>=2.23.0
python-dateutil>=2.8.0
python-dotenv>=1.0.0
